+++ mybatis-plus-core/src/main/java/com/baomidou/mybatisplus/core/conditions/interfaces/Join.java
@@ -96,6 +96,19 @@ public interface Join<This> extends Serializable {
      */
     This apply(boolean condition, String applySql, Object... value);
 
+    /**
+     * 无视优化规则直接拼接到 sql 的最后
+     * 例: last("limit 1")
+     */
+    default This last(String lastSql) {
+        return last(true, lastSql);
+    }
+
+    /**
+     * 无视优化规则直接拼接到 sql 的最后
+     * 例: last("limit 1")
+     */
+    This last(boolean condition, String lastSql);
 
     /**
      * EXISTS ( sql 语句 )
